7th Module Completed!

+++ 10 Py Projects/6-BasicUserInput/1-UserInput.py
@@ -9,5 +9,5 @@ while True:
     if user_input == 'exit':
         break
 
-    temprature = int(user_input)
+    temprature = float(user_input)
     print(weather_condition(temprature))
